requirements.txt
Flask
flask-cors
werkzeug
pandas
psycopg2-binary
SQLAlchemy
python-dotenv
numpy
Jinja2 
pyarrow
